refactor(service-sg-server): fix import path
https://github.com/akky-xxxx/project-my-homepage/issues/44
